Change: Book.bookId (int) -> (long)

+++ app/src/main/java/com/example/bookstory/DAO/Book.java
@@ -9,7 +9,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Book implements Parcelable {
     @PrimaryKey(autoGenerate = true)
-    public int bookId;
+    public long bookId;
 
     public String bookName;
     public int numberOfPages;
@@ -25,7 +25,7 @@ public class Book implements Parcelable {
 
 
     protected Book(Parcel in) {
-        bookId = in.readInt();
+        bookId = in.readLong();
         bookName = in.readString();
         numberOfPages = in.readInt();
         yearOfPublication = in.readInt();
@@ -51,7 +51,7 @@ public class Book implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(bookId);
+        dest.writeLong(bookId);
         dest.writeString(bookName);
         dest.writeInt(numberOfPages);
         dest.writeInt(yearOfPublication);
